Added steak to index page made boilerplate for steak recipe
I modified the index.html page to now also have the
steak recipe webpage I just started, as well as also
adding an unordered list of the links to the recipe pages
to aid with the readability of the index page. I started
the html boilerplate for the steak web page.