Created style sheets for index and recipe pages
Created a style sheet for the index page for the odin recipe project.
Also added the style reference to the html head. In addition I styled the
elements on the index page.

Created the style sheets for the recipe pages as well as referenced the
style sheets in the html head.